fix: hide overflow content for thumbnails
